Refactor `MethodCommand` to `Command`
