+++ src/main/java/dev/scyye/botcommons/commands/Param.java
@@ -1,4 +1,4 @@
-package dev.scyye.botcommons.methodcommands;
+package dev.scyye.botcommons.commands;
 
 import net.dv8tion.jda.api.interactions.commands.OptionType;
