+++ src/main/java/dev/scyye/botcommons/commands/ReplyContext.java
@@ -1,4 +1,4 @@
-package dev.scyye.botcommons.methodcommands;
+package dev.scyye.botcommons.commands;
 
 import dev.scyye.botcommons.menu.MenuManager;
 import lombok.Getter;
